requirements.txt
asgiref==3.7.2
Babel==2.13.1
cobble==0.1.3
Django==4.2.7
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-js-asset==2.1.0
djangorestframework==3.14.0
docxcompose==1.4.0
gunicorn==21.2.0
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
